HTB-12 refactoring code: divide submit handler for sign up
