+++ htb-frontend/src/app/signup/page.tsx
@@ -1,47 +1,13 @@
-'use client'
-
-import { SubmitHandler } from 'react-hook-form';
 import Icon from '../components/icon/Icon';
 import SignUpForm, { SignUpFormValues } from './components/SignUpForm';
 import styles from './page.module.scss'
-import { useState } from 'react';
-import { useRouter } from 'next/navigation';
-
-const signUp = async (username: string, email: string, password: string) => {
-    const response = await fetch('http://localhost:8080/api/auth/signup', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ username, email, password }),
-    });
-
-    if (response.ok) {
-        try {
-            const responseText = await response.text();
-            console.log(responseText);
-        } catch (error) {
-            console.error('Error parsing JSON:', error);
-        }
-    } else {
-        console.error('User registration failed');
-    }
-};
 
 const SignUp: React.FC = () => {
-    const route = useRouter();
-    const [signUpData, setSignUpData] = useState<SignUpFormValues | null>(null);
-
-    const handleSignUp: SubmitHandler<SignUpFormValues> = async (data: SignUpFormValues) => {
-      await signUp(data.username, data.email, data.password).then(() => route.push('/'));
-      setSignUpData(data);
-    };
-
     return (
         <div className={styles.signUpPageWrapper}>
             <div className={styles.signUpWrapper}>
                 <span className={styles.signUpTitle}> 회 원 가 입 </span>
-                <SignUpForm onSignUp={handleSignUp}/>
+                <SignUpForm />
             </div>
             <div className={styles.signUpInfoWrapper}>
                 <Icon iconName='info' iconCustomStyle={{ color: '#FFBB00', marginRight: '10px' }} />
